Create Decision Structures
One cause for speeding is the desire to shorten the time spent traveling. Create a program that calculates the total amount of time saved if you are traveling with an average speed that is above the speed limit as compared to traveling with an average speed exactly at the speed limit. Ask the user for the average speed, speed limit, and distance traveled. Calculate the amount of time saved for the distance traveled.