Rename INPUT OUTPUT PIZZA PROGRAM 1 to Input Output
Assume each pizza has 8 slices. Assuming that there is a family of 4. Ask how many slices everyone in the family eats. Then, calculate how many whole pizzas are required for a family of 4 for one meal. How many are the leftover slices? So, if everyone is eating 3 slices you need 2 pizzas with 4 leftover slices. Do not use loops or lists or functions. In this case, write two programs one with everyone eating the same number of slices (only 1 input statement) and another with everyone eating a different number of slices (two input statements).